rename interface to ITwoPhaseStatefulComponent
