+++ heron/instance/tests/java/org/apache/heron/resource/TestTwoPhaseStatefulBolt.java
@@ -16,45 +16,50 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
 package org.apache.heron.resource;
 
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicBoolean;
 
-import org.apache.heron.api.spout.BaseRichSpout;
-import org.apache.heron.api.spout.SpoutOutputCollector;
+import org.junit.Ignore;
+
+import org.apache.heron.api.bolt.BaseRichBolt;
+import org.apache.heron.api.bolt.OutputCollector;
 import org.apache.heron.api.state.State;
-import org.apache.heron.api.topology.I2PhaseCommitComponent;
+import org.apache.heron.api.topology.ITwoPhaseStatefulComponent;
 import org.apache.heron.api.topology.OutputFieldsDeclarer;
 import org.apache.heron.api.topology.TopologyContext;
+import org.apache.heron.api.tuple.Fields;
+import org.apache.heron.api.tuple.Tuple;
 import org.apache.heron.common.basics.SingletonRegistry;
 
-public class Test2PhaseCommitSpout extends BaseRichSpout
-    implements I2PhaseCommitComponent<String, String> {
+@Ignore
+public class TestTwoPhaseStatefulBolt extends BaseRichBolt
+    implements ITwoPhaseStatefulComponent<String, String> {
+
+  private static final long serialVersionUID = -5160420613503624743L;
+
   @Override
-  public void open(
-      Map<String, Object> conf,
-      TopologyContext context,
-      SpoutOutputCollector collector) {
+  public void prepare(
+      Map<String, Object> map,
+      TopologyContext topologyContext,
+      OutputCollector collector) {
   }
 
   @Override
-  public void nextTuple() {
-    AtomicBoolean shouldStartEmit =
-        (AtomicBoolean) SingletonRegistry.INSTANCE.getSingleton(Constants.SPOUT_SHOULD_START_EMIT);
+  public void execute(Tuple tuple) {
+    CountDownLatch tupleExecutedLatch =
+        (CountDownLatch) SingletonRegistry.INSTANCE.getSingleton(Constants.EXECUTE_LATCH);
 
-    if (shouldStartEmit != null && !shouldStartEmit.get()) {
-      return;
+    if (tupleExecutedLatch != null) {
+      tupleExecutedLatch.countDown();
     }
+  }
 
-    // actually "emit" the tuple
-    CountDownLatch emitLatch =
-        (CountDownLatch) SingletonRegistry.INSTANCE.getSingleton(Constants.EMIT_LATCH);
-
-    if (emitLatch != null) {
-      emitLatch.countDown();
-    }
+  @Override
+  public void declareOutputFields(OutputFieldsDeclarer outputFieldsDeclarer) {
+    outputFieldsDeclarer.declare(new Fields("word"));
   }
 
   @Override
@@ -89,9 +94,5 @@ public class Test2PhaseCommitSpout extends BaseRichSpout
     if (preSaveLatch != null) {
       preSaveLatch.countDown();
     }
-  }
-
-  @Override
-  public void declareOutputFields(OutputFieldsDeclarer declarer) {
   }
 }
